+++ NEWS.md
@@ -5,6 +5,9 @@
 * Added vignette for the logistic regression model (IWLS).
 * Added new references.
 * In `iwls_logit`: renamed variable `mu` to `prob`.
+* Added a Hovmoeller diagram method (`image.foehnix`) with some
+  features. Has to be seen as "under development" at the moment!
+* Removed some warnings/notes.
 
 # foehnix 0.0-5